Add APIs for the cards and the ATM API begin for the cards

+++ server/models/Accounts.php
@@ -52,7 +52,7 @@ class Accounts extends Model {
             `name` VARCHAR(255) NOT NULL,
             `type` INT UNSIGNED NOT NULL,
             `user_id` INT UNSIGNED NOT NULL,
-            `amount` BIGINT UNSIGNED NOT NULL,
+            `amount` DECIMAL(15,2) UNSIGNED NOT NULL,
             `created_at` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
             `updated_at` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
         )');
